bots/tg_calculation_bot/pyproject.toml
[tool.poetry]
name = "tg-calculation"
version = "0.1.0"
description = ""
authors = ["SerghAl <sergey.alabuzhev.personal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.1.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
